requirements.txt
# Framework web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Base de datos
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.25.0

# Configuración
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Validación de email
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Desarrollo
ruff>=0.1.0
mypy>=1.6.0
black>=23.9.0

# Logging y monitoreo
structlog>=23.2.0

# Utilidades
python-dotenv>=1.0.0
